requirements.txt
python-telegram-bot==20.0
requests==2.28.1
watchdog==3.0.0
fastapi==0.95.2
uvicorn==0.22.0